requirements.txt
numpy>=1.23.5,<2.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
flask>=3.0.0
werkzeug>=3.0.0
